src/appli8/requirements.txt
pandas
PyYAML
scikit-learn
python-dotenv
duckdb